Day6

+++ src/components/News/index.js
@@ -10,38 +10,38 @@ import {
 } from '@fortawesome/free-solid-svg-icons'
 
 import * as s from './index.module.scss'
+import dateFormater from '../../utils/dateFormater'
 
 
 export default ({
   handlerRequest,
   deleteRequest,
-  userData,
   data,
   token,
-  loggedIn
+  loggedIn,
+  id,
 }) => {
 
   const match = useRouteMatch()
-  let name = '';
 
-  if (userData.getName) {
-    name = userData.getName()
-  }
+  const tmpEditIcons = (creatorId, i) => {
 
-  const tmpEditIcons = (creatorName, i) => {
+    console.log(creatorId, id)
 
     const _onDelete = () => {
-      deleteRequest({
-        newsId: i._id,
-        token: token,
-      })
+      if (window.confirm('Уверены, что желаете удалить новость?')) {
+        deleteRequest({
+          newsId: i._id,
+          token: token,
+        })
+      }
     }
 
-    if (name && name === creatorName) {
+    if (id && id === creatorId) {
       return (
         <div className={s.icons}>
           <Link to={`${match.path}/${i._id}/edit`}>
-            <button>
+            <button tabIndex='-1'>
               <FontAwesomeIcon icon={faEdit} />
             </button>
           </Link>
@@ -77,19 +77,12 @@ export default ({
             return string
           }
 
-          const date = new Date(i.createDate)
-            .toLocaleDateString('ru', {
-              year: 'numeric',
-              month: 'long',
-              day: 'numeric',
-              hour: 'numeric',
-              minute: 'numeric',
-            })
+          const date = dateFormater(i.createDate)
 
           return (
             <div key={i.createDate} className={s.wrapItem}>
               <div className={s.header}>
-                {tmpEditIcons(i.creator.displayName, i)}
+                {tmpEditIcons(i.creator._id, i)}
               </div>
               <div className={s.body}>
                 <Link
